Renamed the fee rate setting.
The solution suggested in #339 is implemented.  However, I think it would be even better if the code were updated so that the setting was millionths instead of hundred-billionths because millionths is what Bolt11 uses.  This requires a much greater understanding of the code and many more changes than I can make right now.  Some day!  Unless someone else gets to it first :-).